Extract CommandWithArg extractor (#311)
Decouple it from Commands.commands global var
