+++ src/test/scala/me/rexim/morganey/autocompletion/extractors/CommandWithArgExtractorSpec.scala
@@ -1,0 +1,7 @@
+package me.rexim.morganey.autocompletion.extractors
+
+import org.scalatest._
+
+class CommandWithArgExtractorSpec extends FlatSpec with Matchers {
+
+}
